fix(forex): Double -> BigDecimal while parsing OneFrame response

+++ forex-mtl/src/main/scala/forex/services/rates/interpreters/oneframe/Pair.scala
@@ -1,3 +1,3 @@
 package forex.services.rates.interpreters.oneframe
 
-case class Pair(from: String, to: String, bid: Double, ask: Double, price: Double, time_stamp: String)
+case class Pair(from: String, to: String, bid: BigDecimal, ask: BigDecimal, price: BigDecimal, time_stamp: String)
